CSS updates to Dunn for homepage promo boxes
Elsa: Please review the Dunn site performance in all browsers. When we presented the site to the client, it was not displaying properly in Internet Explorer. 
RC: Removed e-mail links. Looking into browser display differences. Resolved homepage promo box issues on Internet Explorer. Header and footer are centered on the browser because a css function that works on Google chrome display: grid does not work on internet explorer, so it overrides to show as mobile view formatting. To correct this display issue, Internet Explorer needs a different css set up, which may require recoding the header and footer.